Resolve #18 - Although not the way I expected
A TODO comment in BoDi's source makes it seem like Factory Registrations
are not cached, but in fact actually they are and they work correctly
as-is.  This TODO can be found in commit:

  a27bb6c0a47981bd44752a3d5be1317d6434e3fb

In file:

  BoDi/BoDi.cs line 404

However, actually there is nothing to do here. BoDi already manages to
cache that object, as proved by the integration test found (in the solution
as it stands as of this commit) at the location:

  BoDi.Tests.RegisterFactoryDelegateTests.ShouldReturnTheSameIfResolvedTwice

So, actually this ticket has been resolved as "no change required".