WIP #25 - Refactor creation of proxying resolvers
This introduces the concept of factory types for resolver proxies, each of
which is used to generate a single proxy instance (or return null, meaning
"no proxy applicable").

Then, the overall resolver factory only needs a collection of proxy
factories.  It loops through them and executes each in turn.

It also makes it much easier to add/change which resolver proxies will be
included, via simple list manipulation.