Unrelated dev - split the resolver factory into a base type and implementation
This makes it clearer as to what part of the class really registers all
of the resolver proxies.  Now all of the supporting logic is pushed
down into the abstract base class.  It's easier to derive from and quicker
to see what needs overriding if you want to change things.